+++ vite.config.ts
@@ -2,6 +2,7 @@ import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
 import { viteSingleFile } from 'vite-plugin-singlefile';
 import path from 'path';
+import pkg from './package.json';
 
 const googleAnalyticsPlugin = {
     name: 'inject-gtag',
@@ -28,6 +29,9 @@ export default defineConfig({
         alias: {
             '@': path.resolve(__dirname, './src'),
         },
+    },
+    define: {
+        __APP_VERSION__: JSON.stringify(pkg.version),
     },
     build: {
         outDir: 'dist-single',